Make ez log a proper module
