+++ modules/log/__main__.py
@@ -1,0 +1,1 @@
+import ez.log as _
